add scene controller to remove scene-specific logic from previewer
