+++ editor/src/scene_viewer/mod.rs
@@ -1,29 +1,18 @@
+use crate::scene::controller::SceneController;
 use crate::{
-    camera::PickingOptions,
     gui::{make_dropdown_list_option, make_dropdown_list_option_with_height},
     load_image,
     message::MessageSender,
     scene_viewer::gizmo::{SceneGizmo, SceneGizmoAction},
     send_sync_message,
-    settings::keys::KeyBindings,
     utils::enable_widget,
-    AddModelCommand, AssetItem, BuildProfile, ChangeSelectionCommand, CommandGroup,
-    DropdownListBuilder, EditorScene, EditorSceneEntry, GraphSelection, InteractionMode, Message,
-    Mode, SaveSceneConfirmationDialogAction, SceneCommand, SceneContainer, Selection,
-    SetMeshTextureCommand, Settings,
+    BuildProfile, DropdownListBuilder, EditorScene, EditorSceneEntry, Message, Mode,
+    SaveSceneConfirmationDialogAction, SceneContainer, Selection, Settings,
 };
 use fyrox::{
-    core::{
-        algebra::{Vector2, Vector3},
-        color::Color,
-        futures::executor::block_on,
-        make_relative_path,
-        math::{plane::Plane, Rect},
-        pool::Handle,
-        uuid::Uuid,
-    },
+    core::{color::Color, math::Rect, pool::Handle, uuid::Uuid},
     engine::Engine,
-    fxhash::{FxHashMap, FxHashSet},
+    fxhash::FxHashMap,
     gui::{
         border::BorderBuilder,
         brush::Brush,
@@ -34,7 +23,7 @@ use fyrox::{
         formatted_text::WrapMode,
         grid::{Column, GridBuilder, Row},
         image::{ImageBuilder, ImageMessage},
-        message::{KeyCode, MessageDirection, MouseButton, UiMessage},
+        message::{MessageDirection, MouseButton, UiMessage},
         stack_panel::StackPanelBuilder,
         tab_control::{
             Tab, TabControl, TabControlBuilder, TabControlMessage, TabDefinition, TabUserData,
@@ -47,31 +36,17 @@ use fyrox::{
         HorizontalAlignment, Orientation, Thickness, UiNode, UserInterface, VerticalAlignment,
         BRUSH_DARKEST,
     },
-    resource::{
-        model::{Model, ModelResourceExtension},
-        texture::{Texture, TextureResource},
-    },
-    scene::{
-        camera::{Camera, Projection},
-        node::Node,
-        Scene,
-    },
+    resource::texture::TextureResource,
+    scene::{camera::Projection, Scene},
     utils::into_gui_texture,
 };
 use std::cmp::Ordering;
 
 mod gizmo;
 
-struct PreviewInstance {
-    instance: Handle<Node>,
-    nodes: FxHashSet<Handle<Node>>,
-}
-
 pub struct SceneViewer {
     frame: Handle<UiNode>,
     window: Handle<UiNode>,
-    pub last_mouse_pos: Option<Vector2<f32>>,
-    pub click_mouse_pos: Option<Vector2<f32>>,
     selection_frame: Handle<UiNode>,
     interaction_modes: FxHashMap<Uuid, Handle<UiNode>>,
     camera_projection: Handle<UiNode>,
@@ -82,7 +57,6 @@ pub struct SceneViewer {
     interaction_mode_panel: Handle<UiNode>,
     contextual_actions: Handle<UiNode>,
     global_position_display: Handle<UiNode>,
-    preview_instance: Option<PreviewInstance>,
     no_scene_reminder: Handle<UiNode>,
     tab_control: Handle<UiNode>,
     scene_gizmo: SceneGizmo,
@@ -322,17 +296,14 @@ impl SceneViewer {
             sender,
             window,
             frame,
-            last_mouse_pos: None,
             interaction_modes: Default::default(),
             selection_frame,
             camera_projection,
-            click_mouse_pos: None,
             play,
             interaction_mode_panel,
             contextual_actions,
             global_position_display,
             build_profile,
-            preview_instance: None,
             stop,
             no_scene_reminder,
             tab_control,
@@ -484,172 +455,44 @@ impl SceneViewer {
 
         if let Some(entry) = scenes.current_scene_entry_mut() {
             let editor_scene = &mut entry.editor_scene;
-            let interaction_mode = entry
-                .current_interaction_mode
-                .and_then(|uuid| entry.interaction_modes.map.get_mut(&uuid));
-
             if let (Some(msg), Mode::Edit) = (message.data::<WidgetMessage>(), mode) {
                 if message.destination() == self.frame() {
+                    let screen_bounds = self.frame_bounds(&engine.user_interface);
                     match *msg {
-                        WidgetMessage::MouseDown { button, pos, .. } => self.on_mouse_down(
-                            button,
-                            pos,
-                            editor_scene,
-                            interaction_mode,
-                            engine,
-                            settings,
-                        ),
-                        WidgetMessage::MouseUp { button, pos, .. } => self.on_mouse_up(
-                            button,
-                            pos,
-                            editor_scene,
-                            interaction_mode,
-                            engine,
-                            settings,
-                        ),
-                        WidgetMessage::MouseWheel { amount, .. } => {
-                            editor_scene.camera_controller.on_mouse_wheel(
-                                amount * settings.camera.zoom_speed,
-                                &mut engine.scenes[editor_scene.scene].graph,
-                                settings,
-                            );
+                        WidgetMessage::MouseDown { button, pos, .. } => {
+                            engine.user_interface.capture_mouse(self.frame());
+
+                            entry.on_mouse_down(button, pos, screen_bounds, engine, settings)
                         }
-                        WidgetMessage::MouseMove { pos, .. } => self.on_mouse_move(
-                            pos,
-                            editor_scene,
-                            interaction_mode,
-                            engine,
-                            settings,
-                        ),
+                        WidgetMessage::MouseUp { button, pos, .. } => {
+                            engine.user_interface.release_mouse_capture();
+
+                            entry.on_mouse_up(button, pos, screen_bounds, engine, settings)
+                        }
+                        WidgetMessage::MouseWheel { amount, .. } => {
+                            entry.on_mouse_wheel(amount, engine, settings);
+                        }
+                        WidgetMessage::MouseMove { pos, .. } => {
+                            entry.on_mouse_move(pos, screen_bounds, engine, settings)
+                        }
                         WidgetMessage::KeyUp(key) => {
-                            if self.on_key_up(
-                                key,
-                                editor_scene,
-                                interaction_mode,
-                                engine,
-                                &settings.key_bindings,
-                            ) {
+                            if entry.on_key_up(key, engine, &settings.key_bindings) {
                                 message.set_handled(true);
                             }
                         }
                         WidgetMessage::KeyDown(key) => {
-                            if self.on_key_down(
-                                key,
-                                editor_scene,
-                                interaction_mode,
-                                engine,
-                                &settings.key_bindings,
-                            ) {
+                            if entry.on_key_down(key, engine, &settings.key_bindings) {
                                 message.set_handled(true);
                             }
                         }
                         WidgetMessage::MouseLeave => {
-                            if let Some(preview) = self.preview_instance.take() {
-                                let scene = &mut engine.scenes[editor_scene.scene];
-
-                                scene.graph.remove_node(preview.instance);
-                            }
+                            entry.on_mouse_leave(engine, settings);
                         }
                         WidgetMessage::DragOver(handle) => {
-                            match self.preview_instance.as_ref() {
-                                None => {
-                                    if let Some(item) =
-                                        engine.user_interface.node(handle).cast::<AssetItem>()
-                                    {
-                                        // Make sure all resources loaded with relative paths only.
-                                        // This will make scenes portable.
-                                        if let Ok(relative_path) = make_relative_path(&item.path) {
-                                            // No model was loaded yet, do it.
-                                            if let Some(model) = engine
-                                                .resource_manager
-                                                .try_request::<Model>(relative_path)
-                                                .and_then(|m| block_on(m).ok())
-                                            {
-                                                let scene = &mut engine.scenes[editor_scene.scene];
-
-                                                // Instantiate the model.
-                                                let instance = model.instantiate(scene);
-
-                                                scene.graph.link_nodes(
-                                                    instance,
-                                                    editor_scene.scene_content_root,
-                                                );
-
-                                                scene.graph[instance]
-                                                    .local_transform_mut()
-                                                    .set_scale(settings.model.instantiation_scale);
-
-                                                let nodes = scene
-                                                    .graph
-                                                    .traverse_handle_iter(instance)
-                                                    .collect::<FxHashSet<Handle<Node>>>();
-
-                                                self.preview_instance =
-                                                    Some(PreviewInstance { instance, nodes });
-                                            }
-                                        }
-                                    }
-                                }
-                                Some(preview) => {
-                                    let screen_bounds = self.frame_bounds(&engine.user_interface);
-                                    let frame_size = screen_bounds.size;
-                                    let cursor_pos = engine.user_interface.cursor_position();
-                                    let rel_pos = cursor_pos - screen_bounds.position;
-                                    let graph = &mut engine.scenes[editor_scene.scene].graph;
-
-                                    let position = if let Some(result) =
-                                        editor_scene.camera_controller.pick(PickingOptions {
-                                            cursor_pos: rel_pos,
-                                            graph,
-                                            editor_objects_root: editor_scene.editor_objects_root,
-                                            scene_content_root: editor_scene.scene_content_root,
-                                            screen_size: frame_size,
-                                            editor_only: false,
-                                            filter: |handle, _| !preview.nodes.contains(&handle),
-                                            ignore_back_faces: settings.selection.ignore_back_faces,
-                                            // We need info only about closest intersection.
-                                            use_picking_loop: false,
-                                            only_meshes: false,
-                                        }) {
-                                        Some(result.position)
-                                    } else {
-                                        // In case of empty space, check intersection with oXZ plane (3D) or oXY (2D).
-                                        let camera = graph[editor_scene.camera_controller.camera]
-                                            .query_component_ref::<Camera>()
-                                            .unwrap();
-
-                                        let normal = match camera.projection() {
-                                            Projection::Perspective(_) => {
-                                                Vector3::new(0.0, 1.0, 0.0)
-                                            }
-                                            Projection::Orthographic(_) => {
-                                                Vector3::new(0.0, 0.0, 1.0)
-                                            }
-                                        };
-
-                                        let plane = Plane::from_normal_and_point(
-                                            &normal,
-                                            &Default::default(),
-                                        )
-                                        .unwrap_or_default();
-
-                                        let ray = camera.make_ray(rel_pos, frame_size);
-
-                                        ray.plane_intersection_point(&plane)
-                                    };
-
-                                    if let Some(position) = position {
-                                        graph[preview.instance].local_transform_mut().set_position(
-                                            settings
-                                                .move_mode_settings
-                                                .try_snap_vector_to_grid(position),
-                                        );
-                                    }
-                                }
-                            }
+                            entry.on_drag_over(handle, screen_bounds, engine, settings);
                         }
                         WidgetMessage::Drop(handle) => {
-                            self.on_drop(handle, engine, editor_scene, settings)
+                            entry.on_drop(handle, screen_bounds, engine, settings);
                         }
                         _ => {}
                     }
@@ -902,223 +745,6 @@ impl SceneViewer {
 
     pub fn frame_bounds(&self, ui: &UserInterface) -> Rect<f32> {
         ui.node(self.frame).screen_bounds()
-    }
-
-    #[must_use]
-    fn on_key_up(
-        &mut self,
-        key: KeyCode,
-        editor_scene: &mut EditorScene,
-        active_interaction_mode: Option<&mut Box<dyn InteractionMode>>,
-        engine: &mut Engine,
-        key_bindings: &KeyBindings,
-    ) -> bool {
-        if editor_scene.camera_controller.on_key_up(key_bindings, key) {
-            return true;
-        }
-
-        if let Some(interaction_mode) = active_interaction_mode {
-            if interaction_mode.on_key_up(key, editor_scene, engine) {
-                return true;
-            }
-        }
-
-        false
-    }
-
-    #[must_use]
-    fn on_key_down(
-        &mut self,
-        key: KeyCode,
-        editor_scene: &mut EditorScene,
-        active_interaction_mode: Option<&mut Box<dyn InteractionMode>>,
-        engine: &mut Engine,
-        key_bindings: &KeyBindings,
-    ) -> bool {
-        if editor_scene
-            .camera_controller
-            .on_key_down(key_bindings, key)
-        {
-            return true;
-        }
-
-        if let Some(interaction_mode) = active_interaction_mode {
-            if interaction_mode.on_key_down(key, editor_scene, engine) {
-                return true;
-            }
-        }
-
-        false
-    }
-
-    fn on_mouse_move(
-        &mut self,
-        pos: Vector2<f32>,
-        editor_scene: &mut EditorScene,
-        active_interaction_mode: Option<&mut Box<dyn InteractionMode>>,
-        engine: &mut Engine,
-        settings: &Settings,
-    ) {
-        let screen_bounds = self.frame_bounds(&engine.user_interface);
-
-        let last_pos = *self.last_mouse_pos.get_or_insert(pos);
-        let mouse_offset = pos - last_pos;
-        editor_scene
-            .camera_controller
-            .on_mouse_move(mouse_offset, &settings.camera);
-        let rel_pos = pos - screen_bounds.position;
-
-        if let Some(interaction_mode) = active_interaction_mode {
-            interaction_mode.on_mouse_move(
-                mouse_offset,
-                rel_pos,
-                editor_scene,
-                engine,
-                screen_bounds.size,
-                settings,
-            );
-        }
-
-        self.last_mouse_pos = Some(pos);
-    }
-
-    fn on_mouse_up(
-        &mut self,
-        button: MouseButton,
-        pos: Vector2<f32>,
-        editor_scene: &mut EditorScene,
-        active_interaction_mode: Option<&mut Box<dyn InteractionMode>>,
-        engine: &mut Engine,
-        settings: &Settings,
-    ) {
-        engine.user_interface.release_mouse_capture();
-
-        let screen_bounds = self.frame_bounds(&engine.user_interface);
-
-        if button == MouseButton::Left {
-            self.click_mouse_pos = None;
-            if let Some(current_im) = active_interaction_mode {
-                let rel_pos = pos - screen_bounds.position;
-                current_im.on_left_mouse_button_up(
-                    editor_scene,
-                    engine,
-                    rel_pos,
-                    screen_bounds.size,
-                    settings,
-                );
-            }
-        }
-
-        editor_scene
-            .camera_controller
-            .on_mouse_button_up(button, &mut engine.scenes[editor_scene.scene].graph);
-    }
-
-    fn on_mouse_down(
-        &mut self,
-        button: MouseButton,
-        pos: Vector2<f32>,
-        editor_scene: &mut EditorScene,
-        active_interaction_mode: Option<&mut Box<dyn InteractionMode>>,
-        engine: &mut Engine,
-        settings: &Settings,
-    ) {
-        engine.user_interface.capture_mouse(self.frame());
-
-        let screen_bounds = self.frame_bounds(&engine.user_interface);
-
-        if button == MouseButton::Left {
-            if let Some(current_im) = active_interaction_mode {
-                let rel_pos = pos - screen_bounds.position;
-
-                self.click_mouse_pos = Some(rel_pos);
-
-                current_im.on_left_mouse_button_down(
-                    editor_scene,
-                    engine,
-                    rel_pos,
-                    screen_bounds.size,
-                    settings,
-                );
-            }
-        }
-
-        editor_scene.camera_controller.on_mouse_button_down(
-            button,
-            engine.user_interface.keyboard_modifiers(),
-            &mut engine.scenes[editor_scene.scene].graph,
-        );
-    }
-
-    fn on_drop(
-        &mut self,
-        handle: Handle<UiNode>,
-        engine: &mut Engine,
-        editor_scene: &mut EditorScene,
-        settings: &Settings,
-    ) {
-        if handle.is_none() {
-            return;
-        }
-
-        let screen_bounds = self.frame_bounds(&engine.user_interface);
-        let frame_size = screen_bounds.size;
-
-        if let Some(item) = engine.user_interface.node(handle).cast::<AssetItem>() {
-            // Make sure all resources loaded with relative paths only.
-            // This will make scenes portable.
-            if let Ok(relative_path) = make_relative_path(&item.path) {
-                if let Some(preview) = self.preview_instance.take() {
-                    let scene = &mut engine.scenes[editor_scene.scene];
-
-                    // Immediately after extract if from the scene to subgraph. This is required to not violate
-                    // the rule of one place of execution, only commands allowed to modify the scene.
-                    let sub_graph = scene.graph.take_reserve_sub_graph(preview.instance);
-
-                    let group = vec![
-                        SceneCommand::new(AddModelCommand::new(sub_graph)),
-                        // We also want to select newly instantiated model.
-                        SceneCommand::new(ChangeSelectionCommand::new(
-                            Selection::Graph(GraphSelection::single_or_empty(preview.instance)),
-                            editor_scene.selection.clone(),
-                        )),
-                    ];
-
-                    self.sender.do_scene_command(CommandGroup::from(group));
-                } else if let Some(tex) = engine
-                    .resource_manager
-                    .try_request::<Texture>(relative_path)
-                    .and_then(|t| block_on(t).ok())
-                {
-                    let cursor_pos = engine.user_interface.cursor_position();
-                    let rel_pos = cursor_pos - screen_bounds.position;
-                    let graph = &engine.scenes[editor_scene.scene].graph;
-                    if let Some(result) = editor_scene.camera_controller.pick(PickingOptions {
-                        cursor_pos: rel_pos,
-                        graph,
-                        editor_objects_root: editor_scene.editor_objects_root,
-                        scene_content_root: editor_scene.scene_content_root,
-                        screen_size: frame_size,
-                        editor_only: false,
-                        filter: |_, _| true,
-                        ignore_back_faces: settings.selection.ignore_back_faces,
-                        use_picking_loop: true,
-                        only_meshes: false,
-                    }) {
-                        let texture = tex.clone();
-                        let mut texture = texture.state();
-                        if texture.data().is_some() {
-                            let node = &mut engine.scenes[editor_scene.scene].graph[result.node];
-
-                            if node.is_mesh() {
-                                self.sender
-                                    .do_scene_command(SetMeshTextureCommand::new(result.node, tex));
-                            }
-                        }
-                    }
-                }
-            }
-        }
     }
 
     pub fn update(&self, editor_scene: &EditorScene, engine: &mut Engine) {
